requirements.txt
numpy
pygame
pyopengl
pyglm
